Moved youtube api related functions into their own file, moved other utility functions into own file
