+++ spotify_api.py
@@ -4,23 +4,18 @@ import base64
 import json
 from requests import get, post
 
+from utils import Utils
+from youtube_api import YoutubeAPI
+
 class SpotifyAPI:
-    
     def __init__(self):
         dotenv.load_dotenv()
         self.client_id = os.getenv("CLIENT_ID")
         self.client_secret = os.getenv("CLIENT_SECRET")
         self.user_id = os.getenv("USER_ID")
         self.downloads_dir = os.getenv("DOWNLOADS_DIR")
-    
-    def get_existing_tracks(self, playlist_name):
-        existing_tracks = []
-        for filename in os.listdir("./" + self.downloads_dir + "/" + playlist_name):
-            if filename.endswith(".mp3"):
-                existing_tracks.append(filename[:-4])
-        
-        return existing_tracks
-    
+
+
     def get_token(self):
         auth_string = self.client_id + ":" + self.client_secret
         auth_bytes = auth_string.encode("utf-8")
@@ -38,13 +33,15 @@ class SpotifyAPI:
         
         return token
 
+
     def get_auth_header(self, token):
         headers = {
             "Authorization": "Bearer " + token
         }
         
         return headers
-        
+
+
     def get_playlist_response(self, token):
         url = f"https://api.spotify.com/v1/users/{self.user_id}/playlists"
         headers = self.get_auth_header(token)
@@ -53,7 +50,8 @@ class SpotifyAPI:
         playlists = response_json["items"]
         
         return playlists
-    
+
+
     def extract_playlist_ids(self, playlists):
         playlists_name_id = []
     
@@ -64,12 +62,14 @@ class SpotifyAPI:
         
         return playlists_name_id
 
+
     def get_playlists(self, token):
         playlists_response = self.get_playlist_response(token)
         playlists = self.extract_playlist_ids(playlists_response)
         
         return playlists
-    
+
+
     def get_track_response(self, playlist_id, token):
         url = f"https://api.spotify.com/v1/playlists/{playlist_id}/tracks"
         headers = self.get_auth_header(token)
@@ -78,6 +78,7 @@ class SpotifyAPI:
         tracks = response_json["items"]
         
         return tracks
+
 
     def extract_track_details(self, track_response):
         track_details = []
@@ -90,8 +91,8 @@ class SpotifyAPI:
                 image_url = images[0]["url"]
                 track_details.append((track_name, artist, image_url))
         return track_details
-        
-        
+
+
     def download_track_image(self, image_url):
         image_response = None
         
@@ -103,6 +104,7 @@ class SpotifyAPI:
         
         return image_response
 
+
     def get_tracks(self, playlist_id, token, existing_tracks, playlist_name):
         track_response = self.get_track_response(playlist_id, token)
         track_details = self.extract_track_details(track_response)
@@ -113,26 +115,34 @@ class SpotifyAPI:
         
         for track_name, artist, image_url in track_details:
             search_string = track_name + " - " + artist
-            sanitized_track_name = sanitize_filename(search_string)
+            sanitized_track_name = Utils.sanitize_filename(search_string)
 
             if sanitized_track_name in existing_tracks:
                 number_of_skips += 1
-                print(f"Skipping \"{search_string}\" as it already exists.")
+                
+                if download_complete:
+                    print()
+                    download_complete = False
+                
+                skip_string = f"Skipping \"{search_string}\" as it already exists."
+                Utils.console_print(skip_string)
                 continue
 
             image_response = self.download_track_image(image_url)
             if image_response is None:
-                print(f"An error occurred while downloading the album art for \"{search_string}\".")
+                image_download_error_string = f"An error occurred while downloading the album art for \"{search_string}\"."
+                Utils.console_print(image_download_error_string)
                 continue
+            
+            youtube_api = YoutubeAPI()
 
-            video_url = get_video_url(search_string)
+            video_url = youtube_api.get_video_url(search_string)
             if video_url is None:
                 tracks_not_found.append(search_string)
                 continue
 
-            download_song(video_url, sanitized_track_name, playlist_name, image_response.content)
+            youtube_api.download_song(video_url, sanitized_track_name, playlist_name, image_response.content)
             number_of_downloads += 1
 
         return tracks_not_found, number_of_downloads, number_of_skips
-        
-        
+
